+++ packages/core/test/client.spec.ts
@@ -149,9 +149,7 @@ describe("Web3Inbox Core Client", () => {
       expect(Web3InboxClient.clientState.initting).toEqual(true);
 
       await waitForEvent(() => {
-        return (
-          Web3InboxClient.clientState.isReady
-        );
+        return Web3InboxClient.clientState.isReady;
       });
 
       expect(Web3InboxClient.clientState.isReady).toEqual(true);
@@ -305,7 +303,7 @@ describe("Web3Inbox Core Client", () => {
 
         let scopeMap: NotifyClientTypes.NotifySubscription["scope"] = {};
 
-        w3iClient.watchScopeMap((watchedScopeMap) => {
+        w3iClient.watchNotificationTypes((watchedScopeMap) => {
           scopeMap = watchedScopeMap;
         });