+++ packages/react/src/hooks/useNotifications.ts
@@ -1,16 +1,34 @@
 import type { NotifyClientTypes } from "@walletconnect/notify-client";
 import { useEffect, useState } from "react";
-import { ErrorOf, HooksReturn, LoadingOf, SuccessOf } from "../types/hooks";
+import { ErrorOf, HooksReturn, SuccessOf } from "../types/hooks";
 import { useWeb3InboxClient } from "./useWeb3InboxClient";
+import { Web3InboxClient } from "@web3inbox/core";
 
-type UseNotificationsData = NotifyClientTypes.NotifyNotification[];
-type NextPageState = (() => Promise<void>) | undefined;
+const waitFor = async (condition: () => boolean) => {
+  return new Promise<void>((resolve) => {
+    setInterval(() => {
+      if (condition()) {
+        resolve();
+      }
+    }, 100);
+  });
+};
+
+
+type UseNotificationsData = (NotifyClientTypes.NotifyNotification & {
+  read: () => void;
+})[];
+
+type NextPageState = () => Promise<void>;
 type UseNotificationsReturn = HooksReturn<
-  NotifyClientTypes.NotifyNotification[],
+  UseNotificationsData,
   {
     hasMore: boolean;
+    hasMoreUnread: boolean;
     isLoadingNextPage: boolean;
     fetchNextPage: NextPageState;
+    markNotificationsAsRead: Web3InboxClient["markNotificationsAsRead"];
+    markAllNotificationsAsRead: Web3InboxClient["markAllNotificationsAsRead"];
   }
 >;
 
@@ -21,38 +39,52 @@ type UseNotificationsReturn = HooksReturn<
  * @param {boolean} [isInfiniteScroll] - Whether or not to keep old notifications in the return array or just the current page
  * @param {string} [account] - Account to get subscriptions notifications from, defaulted to current account
  * @param {string} [domain] - Domain to get subscription notifications from, defaulted to one set in init.
+ * @param {boolean} [unreadFirst] - Should unread notifications be ordered on top regardless of recency
  */
 export const useNotifications = (
   notificationsPerPage: number,
   isInfiniteScroll?: boolean,
   account?: string,
-  domain?: string
+  domain?: string,
+  unreadFirst: boolean = true
 ): UseNotificationsReturn => {
   const { data: w3iClient } = useWeb3InboxClient();
 
-  const [nextPage, setNextPage] = useState<NextPageState>(undefined);
-  const [data, setData] = useState<UseNotificationsData>([]);
+  const [nextPage, setNextPage] = useState<NextPageState>();
+
+  // Using a wrapper because `pageNotifications` delivers the same object reference on update
+  // so `setDataWrapper` won't trigger `useEffect`s to update when it is called. This is because
+  // it is receiving the same valtio object. To mitigate this we wrap data in an object
+  // and create a new object with the data. The alternative would have been to spread the array
+  // coming from pageNotifications but that is unnecessarily inefficient. 
+  const [dataWrapper, setDataWrapper] = useState<{data: UseNotificationsData}>({data: []});
+
   const [isLoadingNextPage, setIsLoadingNextPage] = useState<boolean>(false);
   const [errorNextPage, setErrorNextPage] = useState<string | null>(null);
   const [hasMore, setHasMore] = useState<boolean>(false);
+  const [hasMoreUnread, setHasMoreUnread] = useState<boolean>(false);
   const [error, setError] = useState<null | string>(null);
 
   useEffect(() => {
     if (!w3iClient) return;
 
     try {
+      setIsLoadingNextPage(true);
       const { nextPage: nextPageFunc, stopWatchingNotifications } =
         w3iClient.pageNotifications(
           notificationsPerPage,
           isInfiniteScroll,
           account,
-          domain
-        )((data) => {
-          setData(data.notifications);
-          setHasMore(data.hasMore);
+          domain,
+	  unreadFirst,
+        )((newData) => {
+          setDataWrapper({data: newData.notifications});
+          setIsLoadingNextPage(false);
+          setHasMore(newData.hasMore);
+          setHasMoreUnread(newData.hasMoreUnread);
         });
 
-      setNextPage(nextPageFunc);
+      setNextPage(() => nextPageFunc);
 
       return () => {
         stopWatchingNotifications();
@@ -60,20 +92,22 @@ export const useNotifications = (
     } catch (e: any) {
       setError(e.message);
     }
-  }, [account, domain, notificationsPerPage, isInfiniteScroll, w3iClient]);
+  }, [account, domain, notificationsPerPage, isInfiniteScroll, w3iClient, setDataWrapper]);
 
   const fetchNextPage = async () => {
     setErrorNextPage(null);
     setIsLoadingNextPage(true);
 
-    return new Promise(async (resolve, reject) => {
+    return new Promise(async (resolve) => {
+      // wait for the next page function to be ready
       if (!nextPage) {
-        const err = new Error("No more pages to fetch");
-        setErrorNextPage(err.message);
-        return reject(err);
+        await waitFor(() => Boolean(nextPage));
       }
 
-      return await nextPage()
+      // It is now guaranteed to be truthy.
+      const nextPageFunc = nextPage as NextPageState;
+
+      return await nextPageFunc()
         .then((res) => {
           resolve(res);
           setIsLoadingNextPage(false);
@@ -85,34 +119,89 @@ export const useNotifications = (
     });
   };
 
+  const markNotificationsAsRead = async (notificationIds: string[]) => {
+    await waitFor(() => Boolean(w3iClient));
+    const w3iClientTruthy = w3iClient as Web3InboxClient;
+
+    w3iClientTruthy.markNotificationsAsRead(notificationIds, account, domain)
+
+    // Optimistic data updates
+    setDataWrapper(({data: notifications}) => ({
+      data: notifications.map((notification) => {
+        if (notificationIds.includes(notification.id)) {
+          return {
+            ...notification,
+            isRead: true,
+          };
+        } else {
+          return notification;
+        }
+      })
+    }));
+  };
+
+  const markAllNotificationsAsRead = async () => {
+    await waitFor(() => Boolean(w3iClient));
+    const w3iClientTruthy = w3iClient as Web3InboxClient;
+
+    w3iClientTruthy
+      .markAllNotificationsAsRead(account, domain)
+      .catch(setError)
+      .then(() => {
+        setDataWrapper(({data: notifications}) =>
+	  ({
+	    data: notifications.map((notification) => ({
+            ...notification,
+            isRead: true,
+          }))
+	  })
+        );
+      });
+  };
+
+  // If the domain of the account change, all previous data is invalidated.
+  useEffect(() => {
+    setDataWrapper({data: []});
+    setHasMore(false);
+  }, [domain, account]);
+
   if (isLoadingNextPage) {
     return {
-      data,
+      data: dataWrapper.data,
       hasMore,
+      hasMoreUnread,
       error: null,
       isLoading: false,
       isLoadingNextPage: isLoadingNextPage,
       fetchNextPage,
+      markNotificationsAsRead,
+      markAllNotificationsAsRead,
     } as SuccessOf<UseNotificationsReturn>;
   }
 
   if (errorNextPage || error) {
     return {
-      data,
+      data: dataWrapper.data,
       hasMore,
+      hasMoreUnread,
       error: errorNextPage ?? error,
       isLoading: false,
       isLoadingNextPage: false,
       fetchNextPage,
+      markNotificationsAsRead,
+      markAllNotificationsAsRead,
     } as ErrorOf<UseNotificationsReturn>;
   }
 
   return {
-    data,
+    data: dataWrapper.data,
     hasMore,
+    hasMoreUnread,
     error: null,
     isLoading: false,
     isLoadingNextPage: false,
     fetchNextPage,
+    markNotificationsAsRead,
+    markAllNotificationsAsRead,
   } as SuccessOf<UseNotificationsReturn>;
 };